+++ src/test/java/org/opensearch/security/dlic/rest/api/RollbackVersionApiActionValidationTest.java
@@ -39,8 +39,7 @@ public class RollbackVersionApiActionValidationTest extends AbstractApiActionVal
             null,
             configurationRepository,
             null,
-            null,
-            restApiAdminPrivilegesEvaluator,
+            restApiAuthorizationEvaluator,
             null,
             settings
         );
